+++ src/graphql/event.rs
@@ -6,13 +6,16 @@ mod group;
 mod http;
 mod kerberos;
 mod ldap;
+mod log;
 mod mqtt;
+mod network;
 mod nfs;
 mod ntlm;
 mod rdp;
 mod smb;
 mod smtp;
 mod ssh;
+mod sysmon;
 mod tls;
 
 pub(super) use self::group::EventGroupQuery;
@@ -22,9 +25,10 @@ use self::{
     dns::DnsCovertChannel, ftp::BlockListFtp, ftp::FtpBruteForce, ftp::FtpPlainText,
     http::BlockListHttp, http::DomainGenerationAlgorithm, http::HttpThreat, http::NonBrowser,
     http::RepeatedHttpSessions, http::TorConnection, kerberos::BlockListKerberos,
-    ldap::BlockListLdap, ldap::LdapBruteForce, ldap::LdapPlainText, mqtt::BlockListMqtt,
-    nfs::BlockListNfs, ntlm::BlockListNtlm, rdp::BlockListRdp, rdp::RdpBruteForce,
-    smb::BlockListSmb, smtp::BlockListSmtp, ssh::BlockListSsh, tls::BlockListTls,
+    ldap::BlockListLdap, ldap::LdapBruteForce, ldap::LdapPlainText, log::ExtraThreat,
+    mqtt::BlockListMqtt, network::NetworkThreat, nfs::BlockListNfs, ntlm::BlockListNtlm,
+    rdp::BlockListRdp, rdp::RdpBruteForce, smb::BlockListSmb, smtp::BlockListSmtp,
+    ssh::BlockListSsh, sysmon::WindowsThreat, tls::BlockListTls,
 };
 use super::{
     customer::{Customer, HostNetworkGroupInput},
@@ -145,6 +149,9 @@ async fn fetch_events(
     let mut block_list_smtp_time = start_time;
     let mut block_list_ssh_time = start_time;
     let mut block_list_tls_time = start_time;
+    let mut windows_threat_time = start_time;
+    let mut network_threat_time = start_time;
+    let mut extra_threat_time = start_time;
 
     loop {
         itv.tick().await;
@@ -179,7 +186,10 @@ async fn fetch_events(
             .min(block_list_smb_time)
             .min(block_list_smtp_time)
             .min(block_list_ssh_time)
-            .min(block_list_tls_time);
+            .min(block_list_tls_time)
+            .min(windows_threat_time)
+            .min(network_threat_time)
+            .min(extra_threat_time);
 
         // Fetch event iterator based on time
         let start = i128::from(start) << 64;
@@ -376,7 +386,24 @@ async fn fetch_events(
                         block_list_tls_time = event_time + ADD_TIME_FOR_NEXT_COMPARE;
                     }
                 }
-                EventKind::Log => continue,
+                EventKind::WindowsThreat => {
+                    if event_time >= windows_threat_time {
+                        tx.unbounded_send(value.into())?;
+                        windows_threat_time = event_time + ADD_TIME_FOR_NEXT_COMPARE;
+                    }
+                }
+                EventKind::NetworkThreat => {
+                    if event_time >= network_threat_time {
+                        tx.unbounded_send(value.into())?;
+                        network_threat_time = event_time + ADD_TIME_FOR_NEXT_COMPARE;
+                    }
+                }
+                EventKind::ExtraThreat => {
+                    if event_time >= extra_threat_time {
+                        tx.unbounded_send(value.into())?;
+                        extra_threat_time = event_time + ADD_TIME_FOR_NEXT_COMPARE;
+                    }
+                }
             }
         }
     }
@@ -500,6 +527,12 @@ enum Event {
     BlockListSsh(BlockListSsh),
 
     BlockListTls(BlockListTls),
+
+    WindowsThreat(WindowsThreat),
+
+    NetworkThreat(NetworkThreat),
+
+    ExtraThreat(ExtraThreat),
 }
 
 impl From<database::Event> for Event {
@@ -543,6 +576,9 @@ impl From<database::Event> for Event {
                 RecordType::Ssh(event) => Event::BlockListSsh(event.into()),
                 RecordType::Tls(event) => Event::BlockListTls(event.into()),
             },
+            database::Event::WindowsThreat(event) => Event::WindowsThreat(event.into()),
+            database::Event::NetworkThreat(event) => Event::NetworkThreat(event.into()),
+            database::Event::ExtraThreat(event) => Event::ExtraThreat(event.into()),
         }
     }
 }