+++ CHANGELOG.md
@@ -5,6 +5,18 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add new `WindowsThreat` event message for Windows sysmon events.
+- Add new `NetworkThreat` event message for network events.
+- Add new `ExtraThreat` event message for misc log events.
+
+### Changed
+
+- Updated review-database to 0.23.0.
+
 ## [0.16.0] - 2024-01-15
 
 ### Added
@@ -359,6 +371,7 @@ across our system.
 
 - An initial version.
 
+[Unreleased]: https://github.com/aicers/review-web/compare/0.16.0...main
 [0.16.0]: https://github.com/aicers/review-web/compare/0.15.0...0.16.0
 [0.15.0]: https://github.com/aicers/review-web/compare/0.14.5...0.15.0
 [0.14.5]: https://github.com/aicers/review-web/compare/0.14.4...0.14.5